+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Support bar copy now driven by meta set in the fundraising panel
+
+### Removed
+
+- Hoisted support bar for fundraisers feature
+
 ## [3.11.0] - 2023-06-09
 
 ### Added