+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [video-caption] shortcode with utility TinyMCE button
+
 ### Changed
 
+- Support bar copy now driven by meta set in the fundraising panel
 - Style to allow the_content <figure>s to be thinner than their container and have the caption fit their width
+
+### Removed
+
+- Hoisted support bar for fundraisers feature
 
 ## [3.11.0] - 2023-06-09