+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Use webp assets for support texture backgrounds
+- Optimise assets for Breaking Britain & Foreign Agent
 - [video-caption] shortcode with utility TinyMCE button
 
 ### Changed